Initial import from private svn repo.
The svn server hasn't run in a long time, so commit history is
unfortunately not included. This is all game code and assets,
pretty much just as submitted in 2007.

The forked version of Box2D, as I recall, was to fix a crash
that occurred when freeing a Box2D object during a collision.